Log error when rollout fails
With "set -e", the script immediately terminates when the "kubectl
rollout status deploy" fails. To be able to log information for the
users, the parameter gets removed.

The error gets handled by checking the return value of "kubectl
rollout status deploy"

Additionally, a hint for the wonderland 2 debug how-to gets added.